==> scripts/08-packages_install.R <==
#### Preamble ####
# Purpose: Install all packages for scripts and paper
# Author: Shaotong Li
# Date: 27 November 2024
# Contact: maxst.li@mail.utoronto.ca
# License: MIT
# Pre-requisites: None
# Any other information needed? Please run this script first.

#### Workspace setup ####

### Install packages for scripts ###
install.packages("tidyverse")
install.packages("lubridate")
install.packages("arrow")
install.packages("tidyr")
install.packages("dplyr")
install.packages("readr")
install.packages("caTools")
install.packages("testthat")
install.packages("here")
install.packages("ggplot2")
install.packages("patchwork")
install.packages("sf")
install.packages("rnaturalearth")
install.packages("rnaturalearthdata")
install.packages("ggthemes")
install.packages("knitr")

### Install packages for paper ###
install.packages("kableExtra")
